+++ sdk/agentserver/azure-ai-agentserver-langgraph/CHANGELOG.md
@@ -1,5 +1,65 @@
 # Release History
 
+
+## 1.0.0b7 (2025-12-05)
+
+### Features Added
+
+- Update response with created_by
+
+### Bugs Fixed
+
+- Fixed error response handling in stream and non-stream modes
+
+## 1.0.0b6 (2025-11-26)
+
+### Feature Added
+
+- Support Agent-framework greater than 251112
+
+
+## 1.0.0b5 (2025-11-16)
+
+### Feature Added
+
+- Support Tools Oauth
+
+### Bugs Fixed
+
+- Fixed streaming generation issues.
+
+
+## 1.0.0b4 (2025-11-13)
+
+### Feature Added
+
+- Adapters support tools
+
+### Bugs Fixed
+
+- Pin azure-ai-projects and azure-ai-agents version to avoid version confliction
+
+
+## 1.0.0b3 (2025-11-11)
+
+### Bugs Fixed
+
+- Fixed Id generator format.
+
+- Fixed trace initialization for agent-framework.
+
+
+## 1.0.0b2 (2025-11-10)
+
+### Bugs Fixed
+
+- Fixed Id generator format.
+
+- Improved stream mode error messsage.
+
+- Updated application insights related configuration environment variables.
+
+
 ## 1.0.0b1 (2025-11-07)
 
 ### Features Added